+++ src/components/overlays/SavePrompt.tsx
@@ -6,7 +6,7 @@ import { FiArrowRight } from 'react-icons/fi';
 import propertyCodeLists from '~/codeLists/propertyCodeLists';
 import NodeMeasurementType from '~/enums/nodeMeasurementType';
 import codeListStore from '~/stores/codeListStore';
-import { toDateString } from '~/utils/dateHelpers';
+import { toDateString } from '~/utils/DateUtils';
 import * as s from './savePrompt.scss';
 
 type Model = 'node' | 'stop' | 'link' | 'route' | 'stopArea' | 'line' | 'routePath' | 'lineHeader';
