+++ src/components/sidebar/lineView/LineHeaderTable.tsx
@@ -17,7 +17,7 @@ import { IMassEditLineHeader, LineHeaderMassEditStore } from '~/stores/lineHeade
 import { LineStore } from '~/stores/lineStore';
 import { LoginStore } from '~/stores/loginStore';
 import { NavigationStore } from '~/stores/navigationStore';
-import { areDatesEqual, toMidnightDate } from '~/utils/dateHelpers';
+import { areDatesEqual, toMidnightDate } from '~/utils/DateUtils';
 import FormValidator from '~/validation/FormValidator';
 import SidebarHeader from '../SidebarHeader';
 import LineHeaderForm from './LineHeaderForm';
