+++ src/services/authService.ts
@@ -3,7 +3,7 @@ import navigator from '~/routing/navigator';
 import QueryParams from '~/routing/queryParams';
 import ErrorStore from '~/stores/errorStore';
 import LoginStore from '~/stores/loginStore';
-import CodeListHelper from '~/utils/CodeListHelper';
+import CodeListUtils from '~/utils/CodeListUtils';
 import HttpUtils from '~/utils/HttpUtils';
 
 export interface IAuthorizationResponse {
@@ -32,11 +32,11 @@ class AuthService {
             if (errorResponse.errorTextKey) {
                 let errorMessage;
                 if (errorResponse.email) {
-                    errorMessage = CodeListHelper.getText(errorResponse.errorTextKey, {
+                    errorMessage = CodeListUtils.getText(errorResponse.errorTextKey, {
                         email: errorResponse.email
                     });
                 } else {
-                    errorMessage = CodeListHelper.getText(errorResponse.errorTextKey);
+                    errorMessage = CodeListUtils.getText(errorResponse.errorTextKey);
                 }
                 ErrorStore.addError(errorMessage);
             }
